requirements.txt
accelerate==0.16.0
antlr4-python3-runtime==4.9.3
bitsandbytes==0.37.0
blis==0.7.7
Bottleneck==1.3.4
brotlipy==0.7.0
catalogue==2.0.7
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cryptography==37.0.1
cymem==2.0.6
docker-pycreds==0.4.0
docopt==0.6.2
en-core-web-sm==3.3.0
et-xmlfile==1.1.0
fairscale==0.4.13
filelock==3.9.0
fsspec==2023.5.0
gitdb==4.0.9
GitPython==3.1.27
gmpy2==2.1.2
huggingface-hub==0.14.1
hydra-core==1.2.0
idna==3.3
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
murmurhash==1.0.7
networkx==2.8.4
numexpr==2.8.1
numpy==1.22.3
omegaconf==2.2.3
openai==0.20.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-stubs==1.2.0.61
pathtools==0.1.2
pathy==0.6.1
peft==0.3.0
Pillow==9.0.1
pip==21.2.4
pipreqs==0.4.11
preshed==3.0.6
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pyarrow==8.0.0
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.27.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
sympy==1.11.1
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.64.0
transformers==4.29.2
triton==2.0.0
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.18
wasabi==0.9.1
wheel==0.37.1
yarg==0.1.9
